refactor: add a base class for ocr controllers
